refactor: Changed QuickFilter to components folder
